Multiple changes. Remove the HOLD flag. (#419)
The "on hold" state is now derived on the fly as it can be dynamic.
If another request comes from a client after an non-idempotent
request, then the client knows what it is doing, and these requests
can be pipelined. In that case remove the flag of non-idempotency
from the preceding request.